make isSingleThreaded groupBy query processing overridable at query time
